fix App::onWindowSize

+++ include/reactive/Timer/CPUTimer.hpp
@@ -10,7 +10,8 @@ public:
 
     auto elapsedInNano() const -> double {
         auto endTime = std::chrono::steady_clock::now();
-        return std::chrono::duration_cast<std::chrono::nanoseconds>(endTime - startTime).count();
+        return static_cast<double>(
+            std::chrono::duration_cast<std::chrono::nanoseconds>(endTime - startTime).count());
     }
 
     auto elapsedInMilli() const -> double { return elapsedInNano() / 1000000.0; }
